- Bugfix: ZonePoint realm checks will once again work with the latest SVN.
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/trunk/GameServer@958 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ packets/Client/168/RegionChangeRequestHandler.cs
@@ -40,7 +40,7 @@ namespace DOL.GS.PacketHandler.Client.v168
 		public int HandlePacket(GameClient client, GSPacketIn packet)
 		{
 			ushort JumpSpotID = packet.ReadShort();
-			ZonePoint zonePoint = (ZonePoint)GameServer.Database.SelectObject(typeof(ZonePoint), "`Id` = '" + JumpSpotID + "' AND (`Realm` = '" + client.Player.Realm + "' OR `Realm` = '0' OR `Realm` = NULL)");
+			ZonePoint zonePoint = (ZonePoint)GameServer.Database.SelectObject(typeof(ZonePoint), "`Id` = '" + JumpSpotID + "' AND (`Realm` = '" + (byte)client.Player.Realm + "' OR `Realm` = '0' OR `Realm` = NULL)");
 
 			if (zonePoint == null)
 			{
